♻️ move convert to pdf to attachment, change scope tokens
