+++ froide/foirequest/api_views/attachment.py
@@ -7,7 +7,11 @@ from django.utils.translation import gettext_lazy as _
 from django_filters import rest_framework as filters
 from drf_spectacular.utils import extend_schema
 from rest_framework import mixins, status, viewsets
+from rest_framework.decorators import action
 from rest_framework.response import Response
+
+from froide.helper.storage import make_unique_filename
+from froide.helper.text_utils import slugify
 
 from ..auth import (
     CreateOnlyWithScopePermission,
@@ -18,8 +22,9 @@ from ..permissions import WriteFoiRequestPermission
 from ..serializers import (
     FoiAttachmentSerializer,
     FoiAttachmentTusSerializer,
+    ImageAttachmentConverterSerializer,
 )
-from ..tasks import move_upload_to_attachment
+from ..tasks import convert_images_to_pdf_api_task, move_upload_to_attachment
 
 User = get_user_model()
 
@@ -50,7 +55,7 @@ class FoiAttachmentViewSet(
         CreateOnlyWithScopePermission,
         WriteFoiRequestPermission,
     ]
-    required_scopes = ["make:message"]
+    required_scopes = ["write:attachment"]
 
     def get_serializer_class(self):
         if self.action == "create":
@@ -116,3 +121,77 @@ class FoiAttachmentViewSet(
 
         self.perform_destroy(instance)
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+    @extend_schema(
+        responses={status.HTTP_201_CREATED: FoiAttachmentSerializer},
+        operation_id="convert_images_to_pdf",
+    )
+    @action(
+        detail=False,
+        methods=["post"],
+        serializer_class=ImageAttachmentConverterSerializer,
+    )
+    def convert_to_pdf(self, request, pk=None):
+        serializer = self.get_serializer(data=request.data)
+
+        if serializer.is_valid():
+            message = serializer.validated_data["message"]
+            foirequest = message.request
+
+            attachments = [i["attachment"] for i in serializer.validated_data["images"]]
+
+            if not all(a.belongs_to == message for a in attachments):
+                return Response(
+                    {
+                        "images": [
+                            _("All attachments need to belong to the provided message.")
+                        ]
+                    },
+                    status=status.HTTP_404_NOT_FOUND,
+                )
+
+            if len(attachments) != len(set(attachments)):
+                return Response(
+                    {"images": [_("Images must not be duplicated.")]},
+                    status=status.HTTP_400_BAD_REQUEST,
+                )
+
+            title = serializer.validated_data["title"]
+            existing_names = {a.name for a in message.attachments}
+
+            name = "{}.pdf".format(slugify(title))
+            name = make_unique_filename(name, existing_names)
+
+            can_approve = not foirequest.not_publishable
+            target = FoiAttachment.objects.create(
+                name=name,
+                belongs_to=message,
+                approved=False,
+                filetype="application/pdf",
+                is_converted=True,
+                can_approve=can_approve,
+            )
+
+            FoiAttachment.objects.filter(id__in=[a.id for a in attachments]).update(
+                converted_id=target.id, can_approve=False, approved=False
+            )
+            instructions = [
+                {"rotate": i["rotate"]} for i in serializer.validated_data["images"]
+            ]
+
+            transaction.on_commit(
+                partial(
+                    convert_images_to_pdf_api_task.delay,
+                    attachments,
+                    target,
+                    instructions,
+                    can_approve=can_approve,
+                )
+            )
+
+            return Response(
+                FoiAttachmentSerializer(target, context={"request": request}).data,
+                status=status.HTTP_201_CREATED,
+            )
+        else:
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
