+++ froide/foirequest/tests/test_api_message.py
@@ -4,7 +4,6 @@ from django.urls import reverse
 import pytest
 
 from froide.foirequest.models.event import FoiEvent
-from froide.foirequest.models.message import MessageKind
 from froide.foirequest.tests import factories
 
 
@@ -229,114 +228,4 @@ def test_auth(client, user):
 
     client.login(email=user2.email, password="froide")
     response = client.get(resource_uri)
-    assert response.status_code == 404
-
-
-@pytest.mark.django_db
-def test_convert_attachment(client: Client, user):
-    request = factories.FoiRequestFactory.create(user=user)
-    message = factories.FoiMessageFactory.create(request=request, kind=MessageKind.POST)
-
-    image1 = factories.FoiAttachmentFactory.create(
-        belongs_to=message, filetype="image/png"
-    )
-    image2 = factories.FoiAttachmentFactory.create(
-        belongs_to=message, filetype="image/png"
-    )
-
-    data = {
-        "title": "test",
-        "images": [
-            {
-                "attachment": reverse(
-                    "api:attachment-detail", kwargs={"pk": image1.pk}
-                ),
-                "rotate": 0,
-            },
-            {
-                "attachment": reverse(
-                    "api:attachment-detail", kwargs={"pk": image2.pk}
-                ),
-                "rotate": 90,
-            },
-        ],
-    }
-
-    # needs to be logged in
-    response = client.post(
-        reverse("api:message-convert-to-pdf", kwargs={"pk": message.pk}),
-        data=data,
-        content_type="application/json",
-    )
-    assert response.status_code == 401
-
-    # wrong user
-    other_user = factories.UserFactory.create()
-    assert client.login(email=other_user.email, password="froide")
-    response = client.post(
-        reverse("api:message-convert-to-pdf", kwargs={"pk": message.pk}),
-        data=data,
-        content_type="application/json",
-    )
-    assert response.status_code == 403
-
-    # everything good
-    assert client.login(email=user.email, password="froide")
-    response = client.post(
-        reverse("api:message-convert-to-pdf", kwargs={"pk": message.pk}),
-        data=data,
-        content_type="application/json",
-    )
-    assert response.status_code == 201
-    data = response.json()
-
-    assert data["name"] == "test.pdf"
-    assert data["filetype"] == "application/pdf"
-    assert data["is_converted"] is True
-
-    image1.refresh_from_db()
-    image2.refresh_from_db()
-    assert image1.converted_id == data["id"]
-    assert image2.converted_id == data["id"]
-
-    # can only convert images from that message
-    not_an_image = factories.FoiAttachmentFactory.create(belongs_to=message)
-    data = {
-        "title": "test",
-        "images": [
-            {
-                "attachment": reverse(
-                    "api:attachment-detail", kwargs={"pk": not_an_image.pk}
-                ),
-                "rotate": 0,
-            }
-        ],
-    }
-
-    response = client.post(
-        reverse("api:message-convert-to-pdf", kwargs={"pk": message.pk}),
-        data=data,
-        content_type="application/json",
-    )
-    assert response.status_code == 400
-
-    # can only convert images from that message
-    not_that_message = factories.FoiAttachmentFactory.create(filetype="image/png")
-    data = {
-        "title": "test",
-        "images": [
-            {
-                "attachment": reverse(
-                    "api:attachment-detail", kwargs={"pk": not_that_message.pk}
-                ),
-                "rotate": 0,
-            }
-        ],
-    }
-
-    response = client.post(
-        reverse("api:message-convert-to-pdf", kwargs={"pk": message.pk}),
-        data=data,
-        content_type="application/json",
-    )
     assert response.status_code == 404
