+++ froide/settings.py
@@ -501,7 +501,8 @@ class Base(Configuration):
             "read:email": _("Read user email"),
             "read:request": _("Read your (private) requests"),
             "make:request": _("Make requests on your behalf"),
-            "create:message": _("Add messages and attachments to user's request"),
+            "write:message": _("Add and edit messages"),
+            "write:attachment": _("Add and edit attachments"),
             "follow:request": _("Follow/Unfollow requests"),
             "read:document": _("Read your (private) documents"),
         },
